requirements.txt
schedule~=1.1.0
tweepy~=3.10.0
requests~=2.26.0
mention_bot~=2.0
pyrebase5==5.0.1